+++ client/web/src/insights/InsightsRouter.tsx
@@ -1,6 +1,11 @@
 import MapSearchIcon from 'mdi-react/MapSearchIcon'
 import React from 'react'
-import { RouteComponentProps, Switch, Route } from 'react-router'
+import { RouteComponentProps, Switch, Route, useRouteMatch } from 'react-router'
+
+import { ExtensionsControllerProps } from '@sourcegraph/shared/src/extensions/controller'
+import { PlatformContextProps } from '@sourcegraph/shared/src/platform/context'
+import { SettingsCascadeProps } from '@sourcegraph/shared/src/settings/settings'
+import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
 
 import { AuthenticatedUser } from '../auth'
 import { withAuthenticatedUser } from '../auth/withAuthenticatedUser'
@@ -8,9 +13,6 @@ import { HeroPage } from '../components/HeroPage'
 import { lazyComponent } from '../util/lazyComponent'
 
 import { CreationRoutes } from './pages/creation/CreationRoutes'
-import { SearchInsightCreationPageProps } from './pages/creation/search-insight/SearchInsightCreationPage'
-import { EditInsightPageProps } from './pages/edit/EditInsightPage'
-import { InsightsPageProps } from './pages/insights/InsightsPage'
 import { getExperimentalFeatures } from './utils/get-experimental-features'
 
 const InsightsLazyPage = lazyComponent(() => import('./pages/insights/InsightsPage'), 'InsightsPage')
@@ -25,10 +27,10 @@ const NotFoundPage: React.FunctionComponent = () => <HeroPage icon={MapSearchIco
  * sub-components withing app tree.
  */
 export interface InsightsRouterProps
-    extends RouteComponentProps,
-        Omit<InsightsPageProps, 'isCreationUIEnabled'>,
-        SearchInsightCreationPageProps,
-        Omit<EditInsightPageProps, 'insightID'> {
+    extends SettingsCascadeProps,
+        PlatformContextProps,
+        TelemetryProps,
+        ExtensionsControllerProps {
     /**
      * Authenticated user info, Used to decide where code insight will appears
      * in personal dashboard (private) or in organisation dashboard (public)
@@ -40,32 +42,38 @@ export interface InsightsRouterProps
  * Main Insight routing component. Main entry point to code insights UI.
  */
 export const InsightsRouter = withAuthenticatedUser<InsightsRouterProps>(props => {
-    const { match, ...outerProps } = props
-    const { codeInsightsDashboards } = getExperimentalFeatures(outerProps.settingsCascade)
+    const { platformContext, settingsCascade, telemetryService, extensionsController, authenticatedUser } = props
+
+    const match = useRouteMatch()
+    const { codeInsightsDashboards } = getExperimentalFeatures(settingsCascade)
 
     return (
         <Switch>
-            <Route render={props => <InsightsLazyPage {...outerProps} {...props} />} path={match.url} exact={true} />
+            <Route path={match.url} exact={true}>
+                <InsightsLazyPage
+                    telemetryService={telemetryService}
+                    platformContext={platformContext}
+                    settingsCascade={settingsCascade}
+                    extensionsController={extensionsController}
+                />
+            </Route>
 
-            <Route
-                path={`${match.url}/create`}
-                render={() => (
-                    <CreationRoutes
-                        platformContext={outerProps.platformContext}
-                        authenticatedUser={outerProps.authenticatedUser}
-                        settingsCascade={outerProps.settingsCascade}
-                        telemetryService={outerProps.telemetryService}
-                    />
-                )}
-            />
+            <Route path={`${match.url}/create`}>
+                <CreationRoutes
+                    platformContext={platformContext}
+                    authenticatedUser={authenticatedUser}
+                    settingsCascade={settingsCascade}
+                    telemetryService={telemetryService}
+                />
+            </Route>
 
             <Route
                 path={`${match.url}/edit/:insightID`}
                 render={(props: RouteComponentProps<{ insightID: string }>) => (
                     <EditInsightLazyPage
-                        platformContext={outerProps.platformContext}
-                        authenticatedUser={outerProps.authenticatedUser}
-                        settingsCascade={outerProps.settingsCascade}
+                        platformContext={platformContext}
+                        authenticatedUser={authenticatedUser}
+                        settingsCascade={settingsCascade}
                         insightID={props.match.params.insightID}
                     />
                 )}
@@ -75,7 +83,12 @@ export const InsightsRouter = withAuthenticatedUser<InsightsRouterProps>(props =
                 <Route
                     path={`${match.url}/dashboard/:dashboardId?`}
                     render={(props: RouteComponentProps<{ dashboardId: string }>) => (
-                        <DashboardsLazyPage dashboardID={props.match.params.dashboardId} {...outerProps} />
+                        <DashboardsLazyPage
+                            telemetryService={telemetryService}
+                            extensionsController={extensionsController}
+                            settingsCascade={settingsCascade}
+                            dashboardID={props.match.params.dashboardId}
+                        />
                     )}
                 />
             )}